+++ src/GUI/IGUI.h
@@ -11,7 +11,7 @@ public:
 
 	// thread safe methods to update GUI
 	virtual void LogMessage(int level, const std::string& msg) = 0;
-	virtual void UpdateInfo(int status, int totalConnections, int uptime, int memoryUsage) = 0;
+	virtual void UpdateInfo(int status, int totalConnections, int uptime, double memoryUsage) = 0;
 	virtual void ShowMessageBox(const std::string& title, const std::string& msg, bool fatalError = false) = 0;
 	virtual void ShowMainWindow() = 0;
